requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-core>=0.3.0
langchain-mcp-adapters>=0.1.0
fastmcp>=0.1.0
web3>=7.0.0
eth-account>=0.13.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.28.0
python-dotenv>=1.0.0
pydantic>=2.9.0
langsmith>=0.1.0
